requirements/offline.txt
wandb
hydra-core
mediapy
torch
torchvision
torchaudio
stable-baselines3[extra]
sb3-contrib
moviepy 
imageio

tensordict
h5py
torchrl

flamingo_pytorch
ftfy 
regex 
transformers
git+https://github.com/openai/CLIP.git
accelerate
